fix: module renamings
