+++ aidial_adapter_bedrock/llm/tools/base.py
