+++ aidial_adapter_bedrock/llm/tools/emulator.py
@@ -4,7 +4,7 @@ from typing import List, Optional
 from pydantic import BaseModel
 
 from aidial_adapter_bedrock.llm.message import BaseMessage
-from aidial_adapter_bedrock.llm.tools_emulation.base import ToolConfig
+from aidial_adapter_bedrock.llm.tools.base import ToolConfig
 from aidial_adapter_bedrock.utils.log_config import bedrock_logger as log
